Perform crossweighting

Perform crossweighting

crossweight(
  network_table,
  matrix,
  meta_data = NULL,
  lag = floor(ncol(matrix)/5),
  min = ceiling(ncol(matrix)/50),
  max = floor(ncol(matrix)/12),
  symmetric_filter = FALSE,
  filter_thresh = 0
)

Arguments

network_table

GRN dataframe, the result of running reconstructargetRN or reconstructargetRN_GENIE3

matrix

genes-by-cells expression matrix

meta_data

result of running findDynGenes

lag

lag window on which to run cross-correlation. Cross-correlaiton computed from -lag to +lag.

min

minimum of weighting window. Edges with offsets (or absolute offsets if symmetric_filter=TRUE) less than min will not be negatively weighted.

max

maximum of weighting window. Edges with offsets (or absolute offsets if symmetric_filter=TRUE) greater than max will have weights set to 0.

symmetric_filter

whether or not to employ a symmetric weight scheme. If true, absolute offset is used in place of offset.

filter_thresh

after crossweighting, edges with weights less than filter_thresh will be set to 0.

Value

network_table with offset and weighted_score added

Examples

if (FALSE) {
data("example_matrix")
network_table <- inferCSN(example_matrix, verbose = TRUE)
network_table_new <- crossweight(
  network_table,
  matrix = t(example_matrix)
)
p1 <- network.heatmap(network_table)
p2 <- network.heatmap(network_table_new[, 1:3])
p1 + p2
}