Find peaks that overlap a specific genomic location

This function from:https://github.com/cole-trapnell-lab/cicero-release/blob/master/R/utils.R#L438

find.overlapping.coordinates(coord_list, coord, maxgap = 0)

Arguments

coord_list

A list of coordinates to be searched for overlap in the form chr_100_2000.

coord

The coordinates that you want to find in the form chr1_100_2000.

maxgap

The maximum distance in base pairs between coord and the coord_list that should count as overlapping. Default is 0.

Value

A character vector of the peaks that overlap coord.

Examples

test_coords <- c(
  "chr18_10025_10225",
  "chr18_10603_11103",
  "chr18_11604_13986",
  "chr18_157883_158536",
  "chr18_217477_218555",
  "chr18_245734_246234"
)
find.overlapping.coordinates(test_coords, "chr18:10,100-1246234")
#> [1] "chr18_10025_10225"   "chr18_10603_11103"   "chr18_11604_13986"  
#> [4] "chr18_157883_158536" "chr18_217477_218555" "chr18_245734_246234"