Fit generalized linear model

Fit generalized linear model

fit_glm(formula, data, family = gaussian, ...)

Arguments

formula

An object of class formula with a symbolic description of the model to be fitted.

data

A data.frame containing the variables in the model.

family

A description of the error distribution and link function to be used in the model. See family for mode details.

...

Other parameters for the model fitting function.

Value

A list with two data frames: gof contains goodness of fit measures of the fit and coefs contains the fitted coefficients.