Fit models for gene expression

Fit models for gene expression

fit_models(object, ...)

# S3 method for CSNObject
fit_models(
  object,
  genes = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  tf_cor = 0.1,
  peak_cor = 0,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("srm", "glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  cores = 1,
  ...
)

Arguments

object

The input data, a csn object.

...

Arguments for other methods

genes

A character vector with the target genes to consider for CSN inference. Takes all VariableFeatures in the object per default.

network_name

network_name.

peak_to_gene_method

Character specifying the method to link peak overlapping motif regions to nearby genes. One of Signac or GREAT.

upstream

Integer defining the distance upstream of the gene to consider as potential regulatory region.

downstream

Integer defining the distance downstream of the gene to consider as potential regulatory region.

extend

Integer defining the distance from the upstream and downstream of the basal regulatory region. Only used of `peak_to_gene_method = 'GREAT'`.

only_tss

Logical. Measure distance from the TSS (TRUE) or from the entire gene body (FALSE).

peak_to_gene_domains

GenomicRanges object with regulatory regions for each gene.

tf_cor

Threshold for TF - target gene correlation.

peak_cor

Threshold for binding peak - target gene correlation.

aggregate_rna_col

aggregate_rna_col

aggregate_peaks_col

aggregate_peaks_col

method

A character string indicating the method to fit the model. * 'srm' - Sparse Regression Model. * 'glm' - Generalized Liner Model with glm. * 'glmnet', 'cv.glmnet' - Regularized Generalized Liner Model with glmnet. * 'brms' - Bayesian Regression Models using brms-package. * 'xgb' - Gradient Boosting Regression using xgboost. * 'bagging_ridge' - Bagging Ridge Regression using scikit-learn via reticulate. * 'bayesian_ridge' - Bayesian Ridge Regression using scikit-learn via reticulate.

interaction_term

The interaction term to use in the model between TF and binding site. * '+' for additive interaction. * ':' for 'multiplicative' interaction. * '*' for crossing interaction, i.e. additive AND 'multiplicative'. For more info, see formula

adjust_method

Method for adjusting p-values.

scale

Logical. Whether to z-transform the expression and accessibility matrices.

verbose

Logical value. Whether to print detailed information.

cores

Number of CPU cores used. Setting to parallelize the computation with foreach.

Value

A CSNObject.