Initiate the CSNObject object

Initiate the CSNObject object

initiate_object(object, ...)

# S4 method for class 'matrix'
initiate_object(object, regulators = character(0), targets = character(0), ...)

# S4 method for class 'Seurat'
initiate_object(
  object,
  celltype = NULL,
  celltype_by = NULL,
  filter_mode = c("variable", "celltype", "unfiltered"),
  filter_by = c("aggregate", "celltype"),
  rna_assay = "RNA",
  rna_min_pct = 0.1,
  rna_logfc_threshold = 0.25,
  rna_test_method = "wilcox",
  n_variable_genes = 2000,
  peak_assay = NULL,
  peak_min_pct = 0.05,
  peak_logfc_threshold = 0.1,
  peak_test_method = "LR",
  n_variable_peaks = "q5",
  regions = NULL,
  exclude_exons = TRUE,
  only_pos = TRUE,
  verbose = TRUE,
  p_value = 0.05,
  ...
)

# S4 method for class 'CSNObject'
initiate_object(object, ...)

Arguments

object

A Seurat object containing gene expression and/or chromatin accessibility data.

...

Additional arguments passed to marker detection functions.

regulators

Regulators to consider for CSN inference.

targets

Targets to consider for CSN inference.

celltype

Selected celltype(s) to analyze. If NULL, all cells will be used.

celltype_by

Column name in Seurat meta.data for cell type information. If NULL, all cells will be treated as one group.

filter_mode

Filter mode for identifying cell-type specific features.

filter_by

Character. When filter_mode is "variable", specify whether to calculate variable features using "aggregate" cells or by "celltype". Default is "aggregate".

rna_assay

A character vector indicating the name of the gene expression assay in the Seurat object.

rna_min_pct

Minimum percentage of cells expressing the gene in either population.

rna_logfc_threshold

Minimum log fold-change threshold for RNA data.

rna_test_method

Statistical test to use for gene markers identification.

n_variable_genes

Number of variable features to select when not comparing between groups.

peak_assay

A character vector indicating the name of the chromatin accessibility assay in the Seurat object. If NULL, only RNA-seq data will be processed.

peak_min_pct

Minimum percentage of cells expressing the peak in either population.

peak_logfc_threshold

Minimum log fold-change threshold for peak data.

peak_test_method

Statistical test to use for peak markers identification.

n_variable_peaks

Number of variable peaks to select ("q5" by default).

regions

Candidate regions to consider for binding site inference. If NULL, all peaks regions are considered. regions could be a GRanges object, or a data frame with three columns: chrom, start, end.

exclude_exons

Logical. Whether to consider exons for binding site inference.

only_pos

Only return positive markers.

verbose

Print progress messages.

p_value

Significance level threshold for filtering features (default: 0.05).

Value

CSNObject object.