Detection of metacells from single-cell gene expression matrix

This function detects metacells from a single-cell gene expression matrix using dimensionality reduction and clustering techniques.

meta_cells(
  matrix,
  genes_use = NULL,
  genes_exclude = NULL,
  var_genes_num = min(1000, nrow(matrix)),
  gamma = 10,
  knn_k = 5,
  do_scale = TRUE,
  pc_num = 25,
  fast_pca = TRUE,
  do_approx = FALSE,
  approx_num = 20000,
  directed = FALSE,
  use_nn2 = TRUE,
  seed = 1,
  cluster_method = "walktrap",
  block_size = 10000,
  weights = NULL,
  do_median_norm = FALSE,
  ...
)

Arguments

matrix

Log-normalized gene expression matrix with rows as genes and columns as cells.

genes_use

A vector of genes used to compute PCA, default is NULL.

genes_exclude

A vector of genes to be excluded when computing PCA, default is NULL.

var_genes_num

If genes_use is not provided, var_genes_num genes with the largest variation are used. Default is min(1000, nrow(matrix)).

gamma

Graining level of data (proportion of number of single cells in the initial dataset to the number of metacells in the final dataset), default is 10.

knn_k

Parameter to compute single-cell kNN network, default is 5.

do_scale

Logical value, default is TRUE, whether to scale gene expression matrix when computing PCA.

pc_num

Number of principal components to use for construction of single-cell kNN network, default is 25.

fast_pca

Logical value, default is TRUE, use irlba as a faster version of prcomp.

do_approx

Logical value, default is FALSE, compute approximate kNN in case of a large dataset (>50'000).

approx_num

Number of cells to subsample for an approximate approach, default is 20000.

directed

Logical value, default is FALSE, whether to use directed graph.

use_nn2

Logical value, default is TRUE, whether to use nn2 function.

seed

Seed to use to subsample cells for an approximate approach, default is 1.

cluster_method

Clustering method to identify metacells, default is walktrap. Available methods are walktrap and louvain (not recommended, gamma is ignored).

block_size

Number of cells to map to the nearest metacell at a time (for approx coarse-graining), default is 10000.

weights

Vector of cell weights (NULL by default), used for computing average gene expression within cluster of metacells.

do_median_norm

Logical value, default is FALSE, whether to normalize by median value.

...

Additional parameters passed to other methods.

Value

A metacell matrix where rows represent metacells and columns represent genes.

References

https://github.com/GfellerLab/SuperCell https://github.com/kuijjerlab/SCORPION

Examples

data("example_matrix")
meta_cells_matrix <- meta_cells(
  example_matrix,
  pc_num = 10,
  fast_pca = FALSE
)
dim(meta_cells_matrix)
#> [1] 500  18
meta_cells_matrix[1:6, 1:6]
#> 6 x 6 Matrix of class "dgeMatrix"
#>             g1        g10        g11        g12        g13        g14
#> [1,] 2.0545057 2.08367091 2.19895110 1.51677313 0.11024463 0.01935760
#> [2,] 2.2880546 2.27451038 2.30015670 2.28168354 1.60320552 0.10528643
#> [3,] 1.8903876 0.02312097 0.01589155 0.02682636 0.02432362 0.02568322
#> [4,] 2.0238664 2.13578014 2.10612123 2.12087076 2.16751773 2.08957642
#> [5,] 0.1361268 2.07312365 2.11331095 2.11336063 2.02598610 2.22708277
#> [6,] 2.1689733 0.01482280 0.01606891 0.01890738 0.02267026 0.01781319