Perform crossweighting

Perform crossweighting

Usage

crossweight(
  grnTab,
  expDat,
  xdyn,
  lag = floor(ncol(expDat)/5),
  min = ceiling(ncol(expDat)/50),
  max = floor(ncol(expDat)/12),
  symmetric_filter = FALSE,
  filter_thresh = 0
)

Arguments

grnTab

GRN dataframe, the result of running reconstructGRN

expDat

Genes-by-cells expression matrix

xdyn

result of running findDynGenes

lag

lag window on which to run cross-correlation. Cross-correlaiton computed from -lag to +lag.

min

minimum of weighting window. Edges with offsets (or absolute offsets if symmetric_filter=TRUE) less than min will not be negatively weighted.

max

maximum of weighting window. Edges with offsets (or absolute offsets if symmetric_filter=TRUE) greater than max will have weights set to 0.

symmetric_filter

whether or not to employ a symmetric weight scheme. If true, absolute offset is used in place of offset.

filter_thresh

after crossweighting, edges with weights less than filter_thresh will be set to 0.

Value

grnTab with offset and weighted_score added