Inferring Cell-Specific Gene Regulatory Network

Inferring Cell-Specific Gene Regulatory Network

inferCSN(object, ...)

# S4 method for matrix
inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  k_folds = NULL,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  verbose = FALSE,
  cores = 1,
  ...
)

# S4 method for data.frame
inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  k_folds = NULL,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  verbose = FALSE,
  cores = 1,
  ...
)

Arguments

object

Input object

...

Arguments passed to other methods

penalty

The type of regularization. This can take either one of the following choices: "L0" and "L0L2". For high-dimensional and sparse data, such as single-cell sequencing data, "L0L2" is more effective.

algorithm

The type of algorithm used to minimize the objective function. Currently "CD" and "CDPSI" are supported. The CDPSI algorithm may yield better results, but it also increases running time.

cross_validation

Check whether cross validation is used.

seed

The seed used in randomly shuffling the data for cross-validation.

n_folds

The number of folds for cross-validation.

k_folds

The number of folds for sample split.

r_threshold

Threshold of R^2.

regulators

Regulator genes.

targets

Target genes.

regulators_num

The number of non-zore coef, this value will affect the final performance. The maximum support size at which to terminate the regularization path. Recommend setting this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small portion of non-zeros.

verbose

Print detailed information.

cores

CPU cores.

Value

A data table of gene-gene regulatory relationship

Examples

library(inferCSN)
data("example_matrix")
weight_table <- inferCSN(example_matrix, verbose = TRUE)
#> Running start for < matrix >.
#> Checking input parameters.
#> All parameters check done.
#> Using 'L0' penalty.
#> Using 1 core.
#> Run done.
head(weight_table)
#>     regulator target     weight
#> 162       g17    g18  0.6777506
#> 145       g18    g17  0.5707733
#> 127       g17    g16  0.4664720
#> 9         g18     g1 -0.4652000
#> 291       g10     g9  0.4565135
#> 55        g13    g12  0.4553549

weight_table <- inferCSN(example_matrix, verbose = TRUE, cores = 2)
#> Running start for < matrix >.
#> Checking input parameters.
#> All parameters check done.
#> Using 'L0' penalty.
#> Using 2 cores.
#> Run done.
head(weight_table)
#>     regulator target     weight
#> 162       g17    g18  0.6777506
#> 145       g18    g17  0.5707733
#> 127       g17    g16  0.4664720
#> 9         g18     g1 -0.4652000
#> 291       g10     g9  0.4565135
#> 55        g13    g12  0.4553549