Inferring Cell-Specific Gene Regulatory Network

Inferring Cell-Specific Gene Regulatory Network

inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  cores = 1,
  verbose = FALSE,
  ...
)

# S4 method for matrix
inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  cores = 1,
  verbose = FALSE,
  ...
)

# S4 method for data.frame
inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  cores = 1,
  verbose = FALSE,
  ...
)

# S4 method for Seurat
inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  cores = 1,
  verbose = FALSE,
  aggregate = TRUE,
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction_name = NULL,
  size_factor_normalize = FALSE,
  genome_info = NULL,
  high_corr_cutoff = NULL,
  low_corr_cutoff = NULL,
  rescued = TRUE,
  ...
)

# S4 method for CSNObject
inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  cores = 1,
  verbose = TRUE,
  genes = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  tf_cor = 0.1,
  peak_cor = 0,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("srm", "glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  alpha = 0.5,
  family = "gaussian",
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  ...
)

# S4 method for CSNObjectList
inferCSN(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  cores = 1,
  verbose = TRUE,
  genes = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  tf_cor = 0.1,
  peak_cor = 0,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("srm", "glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  alpha = 0.5,
  family = "gaussian",
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  clusters = NULL,
  ...
)

Arguments

object

The input data for inferCSN.

penalty

The type of regularization. This can take either one of the following choices: L0 and L0L2. For high-dimensional and sparse data, such as single-cell sequencing data, L0L2 is more effective.

algorithm

The type of algorithm used to minimize the objective function. Currently CD and CDPSI are supported. The CDPSI algorithm may yield better results, but it also increases running time.

cross_validation

Check whether cross validation is used.

seed

The seed used in randomly shuffling the data for cross-validation.

n_folds

The number of folds for cross-validation.

percent_samples

The percent of all samples used for sparse.regression. Default set to 1.

r_threshold

Threshold of \(R^2\) or correlation coefficient.

regulators

A character vector with the regulators to consider for CSN inference.

targets

A character vector with the targets to consider for CSN inference.

regulators_num

The number of non-zore coefficients, this value will affect the final performance. The maximum support size at which to terminate the regularization path. Recommend setting this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small portion of non-zeros.

cores

Number of CPU cores used. Setting to parallelize the computation with foreach.

verbose

Logical value. Whether to print detailed information.

...

Parameters for other methods.

aggregate

Logical, whether to aggregate the data.

k_neigh

Number of cells to be aggregated per cluster.

atacbinary

Logical, whether the aggregated scATAC-seq data need binary

max_overlap

The maximum overlapping ratio of two clusters.

reduction_name

The reduction name of extracting the cell coordinates used for aggregating.

size_factor_normalize

Logical, should accessibility values be normalized by size factor.

genome_info

Genome information.

high_corr_cutoff

atacbinary

low_corr_cutoff

atacbinary

rescued

Logical

genes

A character vector with the target genes to consider for CSN inference. Takes all VariableFeatures in the object per default.

network_name

network_name.

peak_to_gene_method

Character specifying the method to link peak overlapping motif regions to nearby genes. One of Signac or GREAT.

upstream

Integer defining the distance upstream of the gene to consider as potential regulatory region.

downstream

Integer defining the distance downstream of the gene to consider as potential regulatory region.

extend

Integer defining the distance from the upstream and downstream of the basal regulatory region. Only used of `peak_to_gene_method = 'GREAT'`.

only_tss

Logical. Measure distance from the TSS (TRUE) or from the entire gene body (FALSE).

peak_to_gene_domains

GenomicRanges object with regulatory regions for each gene.

tf_cor

Threshold for TF - target gene correlation.

peak_cor

Threshold for binding peak - target gene correlation.

aggregate_rna_col

aggregate_rna_col

aggregate_peaks_col

aggregate_peaks_col

method

A character string indicating the method to fit the model. * 'srm' - Sparse Regression Model. * 'glm' - Generalized Liner Model with glm. * 'glmnet', 'cv.glmnet' - Regularized Generalized Liner Model with glmnet. * 'brms' - Bayesian Regression Models using brms-package. * 'xgb' - Gradient Boosting Regression using xgboost. * 'bagging_ridge' - Bagging Ridge Regression using scikit-learn via reticulate. * 'bayesian_ridge' - Bayesian Ridge Regression using scikit-learn via reticulate.

alpha

The elasticnet mixing parameter. See glmnet for details.

family

A description of the error distribution and link function to be used in the model. See family for mode details.

interaction_term

The interaction term to use in the model between TF and binding site. * '+' for additive interaction. * ':' for 'multiplicative' interaction. * '*' for crossing interaction, i.e. additive AND 'multiplicative'. For more info, see formula

adjust_method

Method for adjusting p-values.

scale

Logical. Whether to z-transform the expression and accessibility matrices.

clusters

clusters.

Value

A data table of gene-gene regulatory relationship

A CSNObject.

A CSNObject_list.

Examples

data("example_matrix")
network_table <- inferCSN(example_matrix, verbose = TRUE)
#> Running start for < matrix >.
#> Checking input parameters.
#> Using `L0` penalty.
#> Run done.
head(network_table)
#>     regulator target     weight
#> 162       g17    g18  0.5443805
#> 9         g18     g1 -0.5390453
#> 109       g16    g15  0.4818454
#> 127       g17    g16  0.4599054
#> 91        g15    g14  0.4516497
#> 73        g14    g13  0.4483912

network_table <- inferCSN(example_matrix, cores = 2)
head(network_table)
#>     regulator target     weight
#> 162       g17    g18  0.5443805
#> 9         g18     g1 -0.5390453
#> 109       g16    g15  0.4818454
#> 127       g17    g16  0.4599054
#> 91        g15    g14  0.4516497
#> 73        g14    g13  0.4483912
if (FALSE) {
data("promoter_regions_hg38")
seurat_object <- inferCSN(
 seurat_object,
 genome_info = promoter_regions_hg38
)
}