Construct network for single target gene

Construct network for single target gene

single_network(
  matrix,
  regulators,
  target,
  cross_validation = FALSE,
  seed = 1,
  penalty = "L0",
  algorithm = "CD",
  regulators_num = (ncol(matrix) - 1),
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  verbose = FALSE,
  ...
)

Arguments

matrix

An expression matrix.

regulators

A character vector with the regulators to consider for CSN inference.

target

The target gene.

cross_validation

Check whether cross validation is used.

seed

The seed used in randomly shuffling the data for cross-validation.

penalty

The type of regularization. This can take either one of the following choices: L0, L0L1 and L0L2. For high-dimensional and sparse data, such as single-cell sequencing data, L0L2 is more effective.

algorithm

The type of algorithm used to minimize the objective function. Currently CD and CDPSI are supported. The CDPSI algorithm may yield better results, but it also increases running time.

regulators_num

The number of non-zore coefficients, this value will affect the final performance. The maximum support size at which to terminate the regularization path. Recommend setting this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small portion of non-zeros.

n_folds

The number of folds for cross-validation.

percent_samples

The percent of all samples used for sparse_regression. Default set to 1.

r_threshold

Threshold of \(R^2\) or correlation coefficient.

verbose

Logical value. Whether to print detailed information.

...

Parameters for other methods.

Value

The weight data table of sub-network

Examples

data("example_matrix")
head(
  single_network(
    example_matrix,
    regulators = colnames(example_matrix),
    target = "g1"
  )
)
#>   regulator target       weight
#> 1       g10     g1  0.005805182
#> 2       g11     g1  0.006012177
#> 3       g12     g1  0.010277177
#> 4       g13     g1 -0.001451393
#> 5       g14     g1 -0.012476081
#> 6       g15     g1  0.016008143

single_network(
  example_matrix,
  regulators = "g1",
  target = "g2"
)
#>   regulator target     weight
#> 1        g1     g2 0.08715829