Safe correlation function which returns a sparse matrix without missing values

Safe correlation function which returns a sparse matrix without missing values

sparse_cor(
  x,
  y = NULL,
  method = "pearson",
  allow_neg = TRUE,
  remove_na = TRUE,
  remove_inf = TRUE,
  ...
)

Arguments

x

Sparse matrix or character vector.

y

Sparse matrix or character vector.

method

Method to use for calculating the correlation coefficient.

allow_neg

Logical. Whether to allow negative values or set them to 0.

remove_na

Logical. Whether to replace NA values with 0.

remove_inf

Logical. Whether to replace infinite values with 1.

...

Other arguments passed to the correlation function.

Value

A correlation matrix.

Examples

m1 <- simulate_sparse_matrix(
  1000, 1000,
  density = 0.05
)
m2 <- simulate_sparse_matrix(
  1000, 500,
  density = 0.05
)

all.equal(
  as.matrix(sparse_cor(m1)),
  cor(as_matrix(m1))
)
#> [1] TRUE
all.equal(
  as.matrix(sparse_cor(m1, m2)),
  cor(as_matrix(m1), as_matrix(m2))
)
#> [1] TRUE

system.time(
  sparse_cor(m1)
)
#>    user  system elapsed 
#>   0.076   0.024   0.101 
system.time(
  cor(as_matrix(m1))
)
#>    user  system elapsed 
#>   1.193   0.000   1.193 
system.time(
  sparse_cor(m1, m2)
)
#>    user  system elapsed 
#>   0.063   0.132   0.049 
system.time(
  cor(as_matrix(m1), as_matrix(m2))
)
#>    user  system elapsed 
#>   1.128   0.000   1.128 

# remove infinite values
m1[sample(1:500, 10)] <- Inf
m2[sample(1:500, 10)] <- Inf

# add missing values
m1[sample(1:500, 10)] <- NA
m2[sample(1:500, 10)] <- NA

sparse_cor(m1, m2)[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>        gene_1       gene_2      gene_3       gene_4       gene_5
#> gene_1      .  .            .           .            .          
#> gene_2      . -0.003526201  0.03513284  0.017012522 -0.025108573
#> gene_3      . -0.036149852 -0.01842075 -0.035091244 -0.034126734
#> gene_4      . -0.009618833 -0.01509936 -0.032335080  0.001989427
#> gene_5      . -0.007101459 -0.01018799  0.003607248 -0.016975775